+++ src/main/java/com/diggydwarff/tobacconistmod/TobacconistCreativeTab.java
@@ -76,6 +76,9 @@ public class TobacconistCreativeTab {
                         output.accept(ModBlocks.WILD_FLOWERING_TOBACCO.get());
 
                         output.accept(ModBlocks.TOBACCO_DRYING_RACK.get());
+                        if (CreateCompat.loaded()) {
+                            output.accept(ModBlocks.INDUSTRIAL_DRYING_RACK.get());
+                        }
                         output.accept(ModBlocks.FLUE_FIREBOX.get());
                         output.accept(ModBlocks.TOBACCO_BARREL.get());